fix(store): Fix addRequestParameter mutaion bug
If a parameter other than the last parameter in request.requestParams
is removed, this bug does not allow another parameter to be added.
E.g if requestParams = [1, 2, 3, 4] and item at index 2 (3) is removed,
no item can be added, but if item at the last index (4) is removed,
another can be added.

This commit fixes that, so at whatever index a parameter is removed,
another can be added.